+++ llama/__init__.py
@@ -2,3 +2,4 @@ from .formatter import * # noqa
 from .model import * # noqa
 from .options import * # noqa
 from .llama_cpp import * # noqa
+from .util import * # noqa