+++ llama/minicpmv.py
@@ -2,7 +2,8 @@ __all__ = [
     'minicpmv_completions',
 ]
 
-from typing import Iterator
+import os
+from typing import Any, Optional, Iterator
 
 from transformers import AutoTokenizer
 
@@ -11,7 +12,7 @@ from .llama_cpp import lib, ffi, lock, llama_context_p, llama_batch, llava_image
 from .context import context_init, context_free
 from .clip import clip_init_context, clip_free_context
 from .sampler import sampler_init, grammar_sampler_init, sampler_free, _common_sampler_sample, _common_sampler_accept
-from .util import _common_token_to_piece
+from .util import _common_token_to_piece, messages_to_prompt_image
 from .formatter import get_tokenizer
 from .llava import _llava_image_embed_make_with_filename, _llava_image_embed_free
 from .clip import _clip_uhd_num_image_embeds_col, _clip_process_eval_image_embed
@@ -71,9 +72,26 @@ def _process_eval_image_embed(ctx_llama: llama_context_p, ctx_clip: clip_ctx_p, 
 
 
 def minicpmv_completions(model: 'Model', model_options: ModelOptions, completions_options: CompletionsOptions) -> Iterator[str]:
-    assert isinstance(completions_options.prompt, str)
-    assert completions_options.messages is None, 'messages are not currently supported'
-    assert isinstance(completions_options.image, str)
+    # either prompt/image or messages, but not both
+    assert (
+        (
+            isinstance(completions_options.prompt, str) and
+            isinstance(completions_options.image, str)
+        ) and not completions_options.messages
+    ) or (
+        not (
+            isinstance(completions_options.prompt, str) and
+            isinstance(completions_options.image, str)
+        ) and completions_options.messages
+    )
+
+    image_file: Optional[Any] = None
+
+    # allow only single message
+    if completions_options.messages:
+        prompt, image_file = messages_to_prompt_image(completions_options.messages)
+        completions_options.prompt = prompt
+        completions_options.image = image_file.name # type: ignore
 
     _model: llama_model_p = model._model
 
@@ -116,6 +134,9 @@ def minicpmv_completions(model: 'Model', model_options: ModelOptions, completion
 
     assert embeds != ffi.NULL
     # print(f'{embeds=}')
+
+    if image_file:
+        os.unlink(image_file.name)
 
     n_past: int = 0
     max_tgt_len: int = 256 if completions_options.predict < 0 else completions_options.predict