+++ llama/qwen2vl.py
@@ -1,6 +1,7 @@
 __all__ = ['qwen2vl_completions']
 
-from typing import Any, Iterator
+import os
+from typing import Any, Optional, Iterator
 
 from transformers import AutoTokenizer
 
@@ -10,9 +11,8 @@ from .context import context_init, context_free
 from .clip import clip_init_context, clip_free_context
 from .llava import _llava_image_embed_free, _llava_image_embed_make_with_filename
 from .sampler import sampler_init, grammar_sampler_init, sampler_free, _common_sampler_sample, _common_sampler_accept
-from .util import _llama_decode, _common_token_to_piece, _zero_array
+from .util import _llama_decode, _common_token_to_piece, _zero_array, messages_to_prompt_image
 from .formatter import get_tokenizer
-
 
 begin_vision_token: str = '<|vision_start|>'
 end_vision_token: str = '<|vision_end|>'
@@ -110,9 +110,26 @@ def _eval_tokens(ctx_llama: llama_context_p, tokens: list[llama_token], n_batch:
 
 
 def qwen2vl_completions(model: 'Model', model_options: ModelOptions, completions_options: CompletionsOptions) -> Iterator[str]:
-    assert isinstance(completions_options.prompt, str)
-    assert completions_options.messages is None, 'messages are not currently supported'
-    assert isinstance(completions_options.image, str)
+    # either prompt/image or messages, but not both
+    assert (
+        (
+            isinstance(completions_options.prompt, str) and
+            isinstance(completions_options.image, str)
+        ) and not completions_options.messages
+    ) or (
+        not (
+            isinstance(completions_options.prompt, str) and
+            isinstance(completions_options.image, str)
+        ) and completions_options.messages
+    )
+
+    image_file: Optional[Any] = None
+
+    # allow only single message
+    if completions_options.messages:
+        prompt, image_file = messages_to_prompt_image(completions_options.messages)
+        completions_options.prompt = prompt
+        completions_options.image = image_file.name # type: ignore
 
     _model: llama_model_p = model._model
 
@@ -155,6 +172,9 @@ def qwen2vl_completions(model: 'Model', model_options: ModelOptions, completions
 
     assert embeds != ffi.NULL
     # print(f'{embeds=}')
+
+    if image_file:
+        os.unlink(image_file.name)
 
     n_past: int = 0
     cur_pos_id: int = 0